Added DataAllocator class and the Text file.
Added the class DataAllocator that distributes the names when given the
number of bytes that can be handled. Unprocessed bytes can be returned
and are redistributed before any new names are given out.

I also added the text file that’s hardcoded into the dataAllocator
class so it’s ready to go.