Almost done
Things to do:
-> Remove ProcessData from the Overhead package in Eclipse, I renamed it
to DataProcessor.
-> Download names.zip from Blackboard, takes a long time there's over
100 million names there I think.
->Don't commit the names.txt to the repo, too big and not necessary,
just put it into your own project folder

There's an issue with Data Allocator. When using the namesshort.txt it's
fine because there isn't that many names and 1 Client gets through the
list very quickly. With the larger names file, the constructor is taking
far too long to complete. Calvin, I'd suggest changing it to just
returning an Object and open the file and then have another method that
is constantly run by the Server that continuously reads the file and
adds the names to the queue, as byte arrays rather than Strings maybe,
and then the getBytes method returns the right amount of names from the
top of the queue so it won't need as much to complete. Send me a message
if you want more info